Fixes and dealing with missing img
If there is no image of a movie/person they will be replaced with a hardcoded option added to img.
Unnecessary files were removed. The class files still stay in case I want to use them but the problem is that I can hardly see the reason why could I do it