Redoing the eventListeners managements
After feedback the eventListeners are added for each button, which makes it more self-descriptive and readable.
The next part of the project is to divide a bit the code so after the data is fetched from the API the function to show it will be called in the evenListener. That with the idea that a function does not call a function according to the feedback. Part of the reasoning is to follow to a better level the principals of functional programing.